requirements.txt
amqp==5.2.0
asttokens==2.4.0
backcall==0.2.0
billiard==4.2.0
biopython==1.81
blinker==1.7.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
ete3==3.1.3
exceptiongroup==1.1.3
executing==2.0.0
Flask==3.0.0
fonttools==4.43.1
idna==3.7
ipykernel==6.25.2
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jupyter_client==8.4.0
jupyter_core==5.4.0
kiwisolver==1.4.5
kombu==5.3.7
mailersend==0.5.6
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.1
packaging==23.2
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
PyMySQL==1.1.0
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.32.3
scipy==1.11.3
seaborn==0.13.0
six==1.16.0
stack-data==0.6.3
tornado==6.3.3
traitlets==5.11.2
tzdata==2023.3
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.8
Werkzeug==3.0.1
